+++ TabletopTweaks/Config/Fixes.cs
@@ -12,6 +12,7 @@ namespace TabletopTweaks.Config {
         public bool FixBackgroundModifiers = true;
         public bool FixShadowSpells = true;
         public bool MetamagicStacking = true;
+        public bool SelectiveMetamagicNonInstantaneous = true;
         public SettingGroup Aeon = new SettingGroup();
         public SettingGroup Azata = new SettingGroup();
         public SettingGroup Lich = new SettingGroup();
@@ -23,6 +24,7 @@ namespace TabletopTweaks.Config {
         public ClassGroup Cavalier = new ClassGroup();
         public ClassGroup Fighter = new ClassGroup();
         public ClassGroup Kineticist = new ClassGroup();
+        public ClassGroup Magus = new ClassGroup();
         public ClassGroup Monk = new ClassGroup();
         public ClassGroup Paladin = new ClassGroup();
         public ClassGroup Ranger = new ClassGroup();
@@ -48,6 +50,7 @@ namespace TabletopTweaks.Config {
             FixMountedLongspearModifer = loadedSettings.FixMountedLongspearModifer;
             FixShadowSpells = loadedSettings.FixShadowSpells;
             MetamagicStacking = loadedSettings.MetamagicStacking;
+            SelectiveMetamagicNonInstantaneous = loadedSettings.SelectiveMetamagicNonInstantaneous;
 
             Aeon.LoadSettingGroup(loadedSettings.Aeon);
             Azata.LoadSettingGroup(loadedSettings.Azata);
@@ -61,6 +64,7 @@ namespace TabletopTweaks.Config {
             Cavalier.LoadClassGroup(loadedSettings.Cavalier);
             Fighter.LoadClassGroup(loadedSettings.Fighter);
             Kineticist.LoadClassGroup(loadedSettings.Kineticist);
+            Magus.LoadClassGroup(loadedSettings.Magus);
             Monk.LoadClassGroup(loadedSettings.Monk);
             Paladin.LoadClassGroup(loadedSettings.Paladin);
             Ranger.LoadClassGroup(loadedSettings.Ranger);